Break cart creation logic into new hook
