+++ packages/peregrine/lib/talons/Header/useCartTrigger.js
@@ -1,13 +1,19 @@
-import { useCallback, useMemo } from 'react';
+import { useCallback, useEffect, useMemo } from 'react';
 import { useCartContext } from '@magento/peregrine/lib/context/cart';
 
 export const useCartTrigger = () => {
-    const [{ details }, { toggleCart }] = useCartContext();
+    const [
+        { cartId, details, isLoading },
+        { getCartDetails, toggleCart }
+    ] = useCartContext();
 
-    // TODO: Update the effect to use graphql and update store with item quantity on mount. Do this when converting getCartDetails to graphql.
-    // useEffect(() => {
-    //     getCartDetails();
-    // }, [getCartDetails]);
+    // If we have a cartId but we haven't gotten details yet and we're not in
+    // the middle of fetching the details, we should go get em.
+    useEffect(() => {
+        if (cartId && typeof details.items_qty === 'undefined' && !isLoading) {
+            getCartDetails();
+        }
+    }, [cartId, details.items_qty, getCartDetails, isLoading]);
 
     const itemCount = useMemo(() => {
         return details.items_qty || 0;
